# Revert "Don't lose traces in Booster"